[SPARK-17677][SQL] Break WindowExec.scala into multiple files
